update_reimburse_lower_limit_to_float

+++ tt_accounting/models/tt_reimburse_commission.py
@@ -9,7 +9,7 @@ _logger = logging.getLogger(__name__)
 class ReimburseCommissionTier(models.Model):
     _name = 'tt.reimburse.commission.tier'
 
-    lower_limit = fields.Integer('Lower Limit', default=0)
+    lower_limit = fields.Float('Lower Limit', default=0)
     rac_amount = fields.Float('Commission Multiplier', default=0.0)
     denominator = fields.Integer('Denominator', default=100)
     rac_preview = fields.Char('Commission Preview', readonly=True, compute='_onchange_rac_denominator')
